requirements.txt
alabaster==0.7.12
amqp==5.0.1
anyjson==0.3.3
apturl==0.5.2
args==0.1.0
asgiref==3.3.0
asn1crypto==0.24.0
awscli==1.18.118
Babel==2.8.0
billiard==3.6.3.0
boto3==1.12.18
botocore==1.17.41
Brlapi==0.6.6
cachetools==4.2.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
clint==0.5.1
colorama==0.4.3
command-not-found==0.3
cryptography==3.3.1
csv-diff==0.6
cupshelpers==1.0
cycler==0.10.0
dateparser==0.7.0
defer==1.0.6
defusedxml==0.5.0
dictdiffer==0.8.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.2
django-celery==3.3.1
django-eventlog==1.1
django-heroku=0.0.0
django-redis==4.12.1
djangorestframework==3.12.2
docutils==0.16
filemagic==1.6
futures==3.1.1
fuzzywuzzy==0.18.0
Geohash==1.0
georaptor==2.0.3
glob2==0.7
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
hashids==1.3.1
httplib2==0.18.1
huey==2.3.0
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==0.9.6
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.2.1
meld==3.18.0
netifaces==0.10.4
numpy==1.19.5
oauth==1.0.1
oauth2client==4.1.3
oauthlib==2.1.0
olefile==0.45.1
packaging==20.3
pandas==1.2.0
paytmchecksum==1.7.0
peewee==3.8.0
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.11.3
proximityhash==1.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.9
pycups==1.9.73
pygeohash==1.2.0
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==20.0.0
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python3-openid==3.1.0
pytz==2020.5
pyxdg==0.25
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
reportlab==3.4.0
requests==2.25.0
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
rsa==4.6
s3==3.0.0
s3transfer==0.3.3
SecretStorage==2.3.1
simpledb==0.4.3
simplejson==3.13.2
six==1.15.0
slackclient==1.1.3
snowballstemmer==2.0.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Sphinx==2.4.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.5
sqlparse==0.4.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.2
usb-creator==0.3.3
vine==5.0.0
virtualenv==15.1.0
wadllib==1.3.2
websocket-client==0.48.0
whitenoise==5.2.0
xkit==0.0.0
xmltodict==0.12.0
zipp==3.4.0
zope.interface==4.3.2
